requirements.txt
openai
python-dotenv
python-docx
argparse
rich
requests
gtts